requirements.txt
streamlit
pandas
numpy
bcrypt
python-docx
lottie
matplotlib
